refactor dropdown and fix linter error
