+++ dashboard/app/views/levels/editors/fields/_weblab2_fields.html.haml
@@ -24,7 +24,7 @@
     - available_ai_tutor_modes = @level.available_ai_tutor_modes ? (@level.available_ai_tutor_modes.is_a?(String) ? JSON.parse(@level.available_ai_tutor_modes) : @level.available_ai_tutor_modes) : ['suggest']
     = f.select :available_ai_tutor_modes, options_for_select(@level.class.ai_tutor_modes, available_ai_tutor_modes), {}, {multiple: true, class: "selectize"}
     %p Set the available AI Tutor modes for the level. The default mode is "Suggest". Use shift-click, ctrl-click, or cmd-click to select multiple options.
-    If multiple modes are available, students can switch between modes in the AI Tutor panel, and the default mode will 
+    If multiple modes are available, students can switch between modes in the AI Tutor panel, and the default mode will
     be the most restrictive mode. The order of modes from most to least restrictive is:
     %ol
       %li
